requirements.txt
aiohappyeyeballs == 2.6.1
aiohttp == 3.12.15
aiosignal == 1.4.0
attrs == 25.3.0
blinker == 1.9.0
certifi == 2025.8.3
charset-normalizer == 3.4.3
click == 8.2.1
colorama == 0.4.6
flask == 3.1.2
frozenlist == 1.7.0
idna == 3.10
itsdangerous == 2.2.0
jinja2 == 3.1.6
markupsafe == 3.0.2
multidict == 6.6.4
mysql-connector-python == 9.4.0
propcache == 0.3.2
py-cord == 2.6.1
python-dotenv == 1.1.1
requests == 2.32.5
urllib3 == 2.5.0
werkzeug == 3.1.3
yarl == 1.20.1
